Compare a dimension count as a signed value
GCC with -Werror=sign-compare rejected comparing the int returned by
extract_n_dims_from_ttype against the size_t n_dims of an Array_t,
breaking the build for every translation unit that includes
gpu_utils.h. Cast the dimension count rather than the return value:
a rank is never negative, so the comparison is unchanged.

+++ src/libasr/codegen/gpu_utils.h
@@ -1728,7 +1728,7 @@ inline ASR::expr_t* gpu_scope_array_shape_source(const ASR::Variable_t *var,
         return nullptr;
     }
     if (ASRUtils::extract_n_dims_from_ttype(ASRUtils::expr_type(src))
-            != arr->n_dims) {
+            != static_cast<int>(arr->n_dims)) {
         return nullptr;
     }
     return src;
